+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/tests/Samples/Readme.cs
@@ -2,8 +2,13 @@
 // Licensed under the MIT License.
 
 using System;
-using Azure.Core;
 using Azure.Core.TestFramework;
+
+#region Snippet:ConversationAnalysisClient_Namespaces
+using Azure.Core;
+using Azure.AI.Language.Conversations;
+using Azure.Identity;
+#endregion
 
 namespace Azure.AI.Language.Conversations.Tests.Samples
 {
@@ -12,8 +17,18 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
         public void CreateConversationAnalysisClient()
         {
             #region Snippet:ConversationAnalysisClient_Create
-            Uri endpoint = new Uri("https://myaccount.api.cognitive.microsoft.com");
+            Uri endpoint = new Uri("https://myaccount.cognitive.microsoft.com");
             AzureKeyCredential credential = new AzureKeyCredential("{api-key}");
+
+            ConversationAnalysisClient client = new ConversationAnalysisClient(endpoint, credential);
+            #endregion
+        }
+
+        public void CreateConversationAnalysisClientWithDefaultAzureCredential()
+        {
+            #region Snippet:ConversationAnalysisClient_CreateWithDefaultAzureCredential
+            Uri endpoint = new Uri("https://myaccount.cognitive.microsoft.com");
+            DefaultAzureCredential credential = new DefaultAzureCredential();
 
             ConversationAnalysisClient client = new ConversationAnalysisClient(endpoint, credential);
             #endregion