+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/tests/Samples/Sample3_AnalyzeConversationWithOptions.cs
@@ -50,6 +50,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
             };
 
             Response response = client.AnalyzeConversation(RequestContent.Create(data));
+            #endregion
 
             using JsonDocument result = JsonDocument.Parse(response.ContentStream);
             JsonElement conversationalTaskResult = result.RootElement;
@@ -76,23 +77,20 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                 Console.WriteLine($"Confidence: {entity.GetProperty("confidenceScore").GetSingle()}");
                 Console.WriteLine();
 
-                if (!entity.TryGetProperty("resolutions", out JsonElement resolutions))
+                if (entity.TryGetProperty("resolutions", out JsonElement resolutions))
                 {
-                    continue;
-                }
-
-                foreach (JsonElement resolution in resolutions.EnumerateArray())
-                {
-                    if (resolution.GetProperty("resolutionKind").GetString() == "DateTimeResolution")
+                    foreach (JsonElement resolution in resolutions.EnumerateArray())
                     {
-                        Console.WriteLine($"Datetime Sub Kind: {resolution.GetProperty("dateTimeSubKind").GetString()}");
-                        Console.WriteLine($"Timex: {resolution.GetProperty("timex").GetString()}");
-                        Console.WriteLine($"Value: {resolution.GetProperty("value").GetString()}");
-                        Console.WriteLine();
+                        if (resolution.GetProperty("resolutionKind").GetString() == "DateTimeResolution")
+                        {
+                            Console.WriteLine($"Datetime Sub Kind: {resolution.GetProperty("dateTimeSubKind").GetString()}");
+                            Console.WriteLine($"Timex: {resolution.GetProperty("timex").GetString()}");
+                            Console.WriteLine($"Value: {resolution.GetProperty("value").GetString()}");
+                            Console.WriteLine();
+                        }
                     }
                 }
             }
-            #endregion
 
             Assert.That(response.Status, Is.EqualTo(200));
             Assert.That(conversationPrediction.GetProperty("topIntent").GetString(), Is.EqualTo("Send"));
@@ -104,13 +102,8 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
         {
             ConversationAnalysisClient client = Client;
 
-            #region Snippet:ConversationAnalysis_AnalyzeConversationWithOptionsAsync
-            string projectName = "Menu";
-            string deploymentName = "production";
-#if !SNIPPET
-            projectName = TestEnvironment.ProjectName;
-            deploymentName = TestEnvironment.DeploymentName;
-#endif
+            string projectName = TestEnvironment.ProjectName;
+            string deploymentName = TestEnvironment.DeploymentName;
 
             var data = new
             {
@@ -135,7 +128,9 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                 kind = "Conversation",
             };
 
+            #region Snippet:ConversationAnalysis_AnalyzeConversationWithOptionsAsync
             Response response = await client.AnalyzeConversationAsync(RequestContent.Create(data));
+            #endregion
 
             using JsonDocument result = await JsonDocument.ParseAsync(response.ContentStream);
             JsonElement conversationalTaskResult = result.RootElement;
@@ -162,23 +157,20 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                 Console.WriteLine($"Confidence: {entity.GetProperty("confidenceScore").GetSingle()}");
                 Console.WriteLine();
 
-                if (!entity.TryGetProperty("resolutions", out JsonElement resolutions))
+                if (entity.TryGetProperty("resolutions", out JsonElement resolutions))
                 {
-                    continue;
-                }
-
-                foreach (JsonElement resolution in resolutions.EnumerateArray())
-                {
-                    if (resolution.GetProperty("resolutionKind").GetString() == "DateTimeResolution")
+                    foreach (JsonElement resolution in resolutions.EnumerateArray())
                     {
-                        Console.WriteLine($"Datetime Sub Kind: {resolution.GetProperty("dateTimeSubKind").GetString()}");
-                        Console.WriteLine($"Timex: {resolution.GetProperty("timex").GetString()}");
-                        Console.WriteLine($"Value: {resolution.GetProperty("value").GetString()}");
-                        Console.WriteLine();
+                        if (resolution.GetProperty("resolutionKind").GetString() == "DateTimeResolution")
+                        {
+                            Console.WriteLine($"Datetime Sub Kind: {resolution.GetProperty("dateTimeSubKind").GetString()}");
+                            Console.WriteLine($"Timex: {resolution.GetProperty("timex").GetString()}");
+                            Console.WriteLine($"Value: {resolution.GetProperty("value").GetString()}");
+                            Console.WriteLine();
+                        }
                     }
                 }
             }
-            #endregion
 
             Assert.That(response.Status, Is.EqualTo(200));
             Assert.That(conversationPrediction.GetProperty("topIntent").GetString(), Is.EqualTo("Send"));