+++ sdk/keyvault/Azure.ResourceManager.KeyVault/tests/VaultOperationsTestsBase.cs
@@ -32,7 +32,7 @@ namespace Azure.ResourceManager.KeyVault.Tests
         public AzureLocation Location { get; set; }
 
         public SubscriptionResource Subscription { get; private set; }
-        public AccessPolicyEntry AccessPolicy { get; internal set; }
+        public VaultAccessPolicy AccessPolicy { get; internal set; }
         public string ResGroupName { get; internal set; }
         public Dictionary<string, string> Tags { get; internal set; }
         public Guid TenantIdGuid { get; internal set; }
@@ -97,14 +97,14 @@ namespace Azure.ResourceManager.KeyVault.Tests
 
             IdentityAccessPermissions permissions = new IdentityAccessPermissions
             {
-                Keys = { new KeyPermission("all") },
-                Secrets = { new SecretPermission("all") },
-                Certificates = { new CertificatePermission("all") },
-                Storage = { new StoragePermission("all") },
+                Keys = { new IdentityAccessKeyPermission("all") },
+                Secrets = { new IdentityAccessSecretPermission("all") },
+                Certificates = { new IdentityAccessCertificatePermission("all") },
+                Storage = { new IdentityAccessStoragePermission("all") },
             };
-            AccessPolicy = new AccessPolicyEntry(TenantIdGuid, ObjectId, permissions);
+            AccessPolicy = new VaultAccessPolicy(TenantIdGuid, ObjectId, permissions);
 
-            VaultProperties = new VaultProperties(TenantIdGuid, new KeyVaultSku(KeyVaultSkuFamily.A, KeyVaultSkuName.Standard));
+            VaultProperties = new VaultProperties(TenantIdGuid, new VaultSku(VaultSkuFamily.A, VaultSkuName.Standard));
 
             VaultProperties.EnabledForDeployment = true;
             VaultProperties.EnabledForDiskEncryption = true;
@@ -112,13 +112,13 @@ namespace Azure.ResourceManager.KeyVault.Tests
             VaultProperties.EnableSoftDelete = true;
             VaultProperties.SoftDeleteRetentionInDays = DefSoftDeleteRetentionInDays;
             VaultProperties.VaultUri = new Uri("http://vaulturi.com");
-            VaultProperties.NetworkAcls = new NetworkRuleSet() {
+            VaultProperties.NetworkRuleSet = new VaultNetworkRuleSet() {
                 Bypass = "AzureServices",
                 DefaultAction = "Allow",
                 IPRules =
                 {
-                    new IPRule("1.2.3.4/32"),
-                    new IPRule("1.0.0.0/25")
+                    new VaultIPRule("1.2.3.4/32"),
+                    new VaultIPRule("1.0.0.0/25")
                 }
             };
             VaultProperties.AccessPolicies.Add(AccessPolicy);
@@ -126,7 +126,7 @@ namespace Azure.ResourceManager.KeyVault.Tests
             ManagedHsmCollection = ResourceGroupResource.GetManagedHsms();
             ManagedHsmProperties = new ManagedHsmProperties();
             ManagedHsmProperties.InitialAdminObjectIds.Add(ObjectId);
-            ManagedHsmProperties.CreateMode = CreateMode.Default;
+            ManagedHsmProperties.CreateMode = ManagedHsmCreateMode.Default;
             ManagedHsmProperties.EnableSoftDelete = true;
             ManagedHsmProperties.SoftDeleteRetentionInDays = DefSoftDeleteRetentionInDays;
             ManagedHsmProperties.EnablePurgeProtection = false;
@@ -135,7 +135,7 @@ namespace Azure.ResourceManager.KeyVault.Tests
                 Bypass = "AzureServices",
                 DefaultAction = "Deny" //Property properties.networkAcls.ipRules is not supported currently and must be set to null.
             };
-            ManagedHsmProperties.PublicNetworkAccess = PublicNetworkAccess.Disabled;
+            ManagedHsmProperties.PublicNetworkAccess = ManagedHsmPublicNetworkAccess.Disabled;
             ManagedHsmProperties.TenantId = TenantIdGuid;
         }