+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/CHANGELOG.md
@@ -4,13 +4,24 @@
 
 ### Features Added
 
-* Added `ConversationAnalysisProjectsClient` to manage projects.
-* Added conversation issue summarization as a long-running operation.
-* Added conversation personally identifiable information (PII) extraction as a long-running operation.
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.0.0 (2022-06-27)
+
+### Features Added
+
+- Added `ConversationAuthoringClient` to manage authoring projects.
+- Added support for Azure Active Directory (AAD) authentication.
 
 ### Breaking Changes
 
+- `ConversationAnalysisClient.AnalyzeConversation` and `AnalyzeConversationAsync` now take a `RequestContent` and `RequestContext` for more control and flexibility.
 - Removed all models. See README.md for samples to use this client library.
+- Renamed `ConversationAnalysisClientOptions` to `ConversationsClientOptions`.
 
 ## 1.0.0-beta.3 (2022-04-20)